+++ src/components/galleryPage/GalleryImage.tsx
