+++ src/components/galleryPage/ImageModal.tsx
