+++ src/components/galleryPage/ModalInput.tsx
